cleaned

+++ include/util.h
@@ -5,6 +5,6 @@
 #define MATLAB_1_2SQRT2 0.707106781186548
 
 using ul = unsigned long;
-using single_p = float;
+using single_p = double;
 using double_p = double;
 #endif //TUMOR_UTIL_H
